refactor(loader): revert over-extracted ep files; keep EP additions inline in upstream files
ep-gen-code-bundle.ts and ep-loader-export-opts.ts required pulling upstream
code (LOADER_CACHE_BUST, makeBundleBucketPath, LOADER_CODEGEN_OPTS_DEFAULTS
etc) into ep-prefixed files to avoid circular deps, which defeats the purpose.

Instead: keep makeExportOpts and tryGetCachedPublishedBundle as clearly-marked
private functions inside gen-code-bundle.ts. Only ep-s3-cache.ts remains — it
contains genuinely standalone new code with no upstream dependencies.
